+++ http-ktor-2/src/jvmMain/kotlin/io/github/bbasinsk/http/ktor2/Example.kt
@@ -1,19 +1,21 @@
 package io.github.bbasinsk.http.ktor2
 
 import io.github.bbasinsk.http.Http
+import io.github.bbasinsk.http.header
 import io.github.bbasinsk.http.openapi.Info
 import io.github.bbasinsk.http.openapi.Server
+import io.github.bbasinsk.http.query
 import io.github.bbasinsk.schema.Schema
 import io.github.bbasinsk.schema.java.instant
 import io.github.bbasinsk.schema.java.localDate
 import io.github.bbasinsk.schema.java.uuid
 import io.github.bbasinsk.schema.transform
 import io.github.bbasinsk.tuple.tupleValues
-import io.ktor.server.cio.CIO
-import io.ktor.server.engine.embeddedServer
+import io.ktor.server.cio.*
+import io.ktor.server.engine.*
 import java.time.Instant
 import java.time.LocalDate
-import java.util.UUID
+import java.util.*
 import kotlin.random.Random