+++ http/src/commonMain/kotlin/io/github/bbasinsk/http/ContentType.kt
@@ -7,8 +7,10 @@ sealed class ContentType(open val contentType: String, open val contentSubtype: 
     data object Json : ContentType("application", "json")
     data object Avro : ContentType("avro", "binary")
     data object Plain : ContentType("text", "plain")
+    data object Html : ContentType("text", "html")
     data object EventStream : ContentType("text", "event-stream")
     data object MultipartFormData : ContentType("multipart", "form-data")
+    data object FormUrlEncoded : ContentType("application", "x-www-form-urlencoded")
 
     sealed class Image(override val contentType: String, override val contentSubtype: String) : ContentType(contentType, contentSubtype) {
         data object Jpeg : Image("image", "jpeg")